% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipr-package.R
\docType{package}
\name{clipr-package}
\alias{clipr-package}
\title{clipr: Read and Write from the System Clipboard}
\description{
Simple utility functions to read from and write to the Windows, OS X, and X11 clipboards.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdlincoln/clipr}
  \item \url{http://matthewlincoln.net/clipr/}
  \item Report bugs at \url{https://github.com/mdlincoln/clipr/issues}
}

}
\author{
\strong{Maintainer}: Matthew Lincoln \email{matthew.d.lincoln@gmail.com} (\href{https://orcid.org/0000-0002-4387-3384}{ORCID})

Other contributors:
\itemize{
  \item Louis Maddox [contributor]
  \item Steve Simpson [contributor]
  \item Jennifer Bryan [contributor]
}

}
\keyword{internal}
