% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipr_addin.R
\name{clipr_addin}
\alias{clipr_addin}
\title{RStudio addin}
\usage{
clipr_addin()
}
\description{
Write the results of the selected expression to the
clipboard
}
\note{
Requires the \href{https://cran.r-project.org/package=rstudioapi}{rstudioapi}
package.
}

