% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipr_addin.R
\name{clipr_output_addin}
\alias{clipr_output_addin}
\title{RStudio addin}
\usage{
clipr_output_addin()
}
\description{
Copies the console output (rather than the value) of a selected expression to
the system clipboard
}
\note{
Requires the
  \href{https://cran.r-project.org/package=rstudioapi}{rstudioapi} package.
}

