% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_data.R
\name{copy_data.default}
\alias{copy_data.default}
\title{Copy data from R - default}
\usage{
\method{copy_data}{default}(obj, row_names = FALSE, max_size = 2^20)
}
\arguments{
\item{obj}{data of type \code{data.frame}, \code{zoo}, \code{xts} or \code{List}}

\item{row_names}{include row names - default FALSE}

\item{max_size}{specify max directory size - default 2^20 bytes}
}
\description{
Copy data from R - default
}
