% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_data.R
\name{copy_data}
\alias{copy_data}
\title{Copy data from R}
\usage{
copy_data(obj, row_names = FALSE, max_size = 2^20)
}
\arguments{
\item{obj}{data of type \code{data.frame}, \code{zoo}, \code{xts} or \code{List}}

\item{row_names}{include row names - default FALSE}

\item{max_size}{specify max directory size - default 2^20 bytes}
}
\value{
nothing. data attached to clipboard
}
\description{
Copy data from R
}
