% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_graphic.R
\name{copy_graphic}
\alias{copy_graphic}
\title{copy graphics generated in R to clipboard}
\usage{
copy_graphic(graphic, size = c("sar", "wide", "custom"), ...)
}
\arguments{
\item{graphic}{graphic generated in R}

\item{size}{string indicating the size (sar, wide or custom) of the new graphics device to copy to. defaults to standard aspect ratio}

\item{...}{additional arguments to pass to the graphics device such as width, height and mar}
}
\value{
nothing. graphic is attached to clipboard
}
\description{
copy graphics generated in R to clipboard
}
