% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_paste_table.R
\name{copy_table}
\alias{copy_table}
\title{Copy a tabular object}
\usage{
copy_table(x = .Last.value, quote = FALSE, sep = "\\t", na = "",
  row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{A tabular object to be copied. This can be \code{data.frame},
\code{table}, \code{matrix} or other tabular classes. The default is
\code{.Last.Value}. See details for a recommended workflow.}

\item{quote}{a logical value (\code{TRUE} or \code{FALSE}) or a
    numeric vector.  If \code{TRUE}, any character or factor columns
    will be surrounded by double quotes.  If a numeric vector, its
    elements are taken as the indices of columns to quote.  In both
    cases, row and column names are quoted if they are written.  If
    \code{FALSE}, nothing is quoted.}

\item{sep}{the field separator string.  Values within each row of
    \code{x} are separated by this string.}

\item{na}{the string to use for missing values in the data.}

\item{row.names}{either a logical value indicating whether the row
    names of \code{x} are to be written along with \code{x}, or a
    character vector of row names to be written.}

\item{col.names}{either a logical value indicating whether the column
    names of \code{x} are to be written along with \code{x}, or a
    character vector of column names to be written.  See the section on
    \sQuote{CSV files} for the meaning of \code{col.names = NA}.}

\item{...}{Further parameters passed to \code{\link{function}}.}
}
\description{
This function allows users to copy a tabular object (e.g. \code{data.frame},
\code{table}, etc) in R so they can paste it to other software.
}
\details{
Since the function is called for its side effect (writing to clipboard), it
always silently return the \code{x} argument. This allows it to be used
in piped operations.

This function is designed to be used interactively in the console. The
default values for the \code{x} argument is \code{.Last.Value}, which is
an internal R object that stores the value of the last evaluated expression.
This allows users to copy the last printed results by simply typing
\code{copy_table()}.

This function also takes explicit arguments for writing tables instead of
taking dots (\code{...}). This allows auto-completion of commonly used
argument names, like \code{na}, or \code{col.names}.

The combination of these features enables this function to be used in such
a workflow:
\enumerate{
  \item Send script to console or type commands in console until
  satisfied with a table output (\code{data.frame}, \code{table}, etc).
  \item Call \code{copy_table()} in console to copy the output.
  \item Paste it somewhere else, usually a spreadsheet software like Excel.
  \item If not satisfied with the output after pasting, come back to R and
  re-exexcute \code{copy_table()} with additional
  arguments, like \code{na = "."}.
}

In the last step, the function can be called again with minimal additional
typing because the table object is stored in \code{.Last.Value} after the
first call to \code{copy_table()}, which is then used for the second call by
default.
}
\examples{
\dontrun{
# if interactively in the console ...
summary(mtcars)
copy_table()  # go paste it
copy_table(col.names = FALSE)  # try paste it again

# copy tables explicitly
obj <- summary(mtcars)
copy_table(obj)
}

}
