% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diffCLIP_BUM_HMM.R
\name{diffCLIP_BUM_HMM}
\alias{diffCLIP_BUM_HMM}
\title{Function BUM_FWBW}
\usage{
diffCLIP_BUM_HMM(filename, Nc, Nt)
}
\arguments{
\item{filename}{a matrix containing the number of mapped reads in each location, for each sample.}

\item{Nc}{number of replicates in condition 1 (or control)}

\item{Nt}{number of replicates in condition 2 (or treatment)}
}
\value{
a matrix containing posteriors in Beta-Uniform mixture HMM
}
\description{
diffCLIP_BUM_HMM compares CLIP data using HMM with Beta-Uniform Mixture emission (BUM model). Takes as input a matrix, two integers (number of replicates in condition 1 and condition 2 respectively), and generates a vector containing all the log ratios in read counts between replicates at every location.
}

