% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_type.R
\name{is_clipr_available}
\alias{is_clipr_available}
\title{Is the system clipboard available?}
\usage{
is_clipr_available()
}
\value{
boolean
}
\description{
Checks to see if the system clipboard is writeable/readable.
}

