% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_df.R
\name{paste_df}
\alias{paste_df}
\title{Paste object from clipboard}
\usage{
paste_df()
}
\value{
\code{data.frame} of the data on the clipboard.
}
\description{
Returns data on clipboard as a \code{data.frame}.
}
