% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clip_tbl.R
\name{read_clip_tbl}
\alias{read_clip_tbl}
\title{Transforms Output of read_clip into dataframe.}
\usage{
read_clip_tbl(x = read_clip(), ...)
}
\arguments{
\item{x}{Defaults to reading from the clipboard, but can be substituted by a
character vector already generated by \code{\link{read_clip}}.}

\item{\ldots}{Options to pass to \code{link[readr]{read_tsv}}}
}
\value{
A data.frame with the contents of the clipboard. If the system
  clipboard is empty, returns \code{NULL}
}
\description{
Transforms clipped content into a data.frame by putting
\code{\link{read_clip}} output by using \code{link[readr]{read_tsv}}.
}

