% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clipboard.R
\name{write_clip}
\alias{write_clip}
\title{Write clipboard}
\usage{
write_clip(content)
}
\arguments{
\item{content}{A character vector to be written to the system clipboard}
}
\value{
On successfully writing the input to the clipboard, this function
  returns the same input for use in piped operations.
}
\description{
Write a character vector to the system clipboard
}

