% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{write_last_clip}
\alias{write_last_clip}
\title{Write contents of the last R expression to the clipboard}
\usage{
write_last_clip(...)
}
\arguments{
\item{\ldots}{Pass other options to \code{\link[=write_clip]{write_clip()}}.}
}
\description{
Write contents of the last R expression to the clipboard
}
\note{
This is a wrapper function for \code{write_clip(.Last.value)}
}
