% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl2df_clip.R
\name{xl2df_clip}
\alias{xl2df_clip}
\title{Transforms Output of Read_clip into Data.Frame}
\usage{
xl2df_clip(clip = clipr::read_clip(), headers = TRUE,
  stringsAsFactor = FALSE)
}
\arguments{
\item{clip}{the excel table clip to transform to data.frame}

\item{headers}{logical for if clip contains colnames or a vector of colnames to use}

\item{stringsAsFactor}{logical for using stringAsFactors, default is FALSE}
}
\description{
Transforms Output of Read_clip into Data.Frame
}
\examples{
# excel contents
#' #            thingOne        thingTwo
# 1 the thing I might the thing I may
# 2    become despite coding this way

# excel contents after read_clip copies
clip_contents_from_excel_table <- c(
"thingOne\\tthingTwo",
"the thing I might\\tthe thing I may",
"become despite\\tcoding this way")

cat_in_the_hat <- xl2df_clip(clip_contents_from_excel_table )
print(cat_in_the_hat)
#            thingOne        thingTwo
# 1 the thing I might the thing I may
# 2    become despite coding this way
}

