% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clear_clip}
\alias{clear_clip}
\title{Clear clipboard}
\usage{
clear_clip(...)
}
\arguments{
\item{\ldots}{Pass other options to \link{write_clip}.}
}
\description{
Clear the system clipboard.
}
\note{
This is a simple wrapper function for write_clip("")
}
