% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_type.R
\name{is_clipr_available}
\alias{is_clipr_available}
\title{Is the system clipboard available?}
\usage{
is_clipr_available()
}
\value{
boolean
}
\description{
Checks to see if the system clipboard is writeable/readable. This may be
useful if you are developing a package that relies on \link{clipr} and need
to ensure that it will skip tests on machines (e.g. CRAN, Travis) where
the system clipboard may not be available.
}
\examples{
\dontrun{
# When using testthat:
library(testthat)
skip_if_not(is_clipr_available())
}

}

