% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clipboard.R
\name{read_clip}
\alias{read_clip}
\title{Read clipboard}
\usage{
read_clip()
}
\value{
A character vector with the contents of the clipboard. If the system
  clipboard is empty, returns NULL
}
\description{
Read the contents of the system clipboard into a character vector.
}
\examples{
clip_text <- read_clip()
}

