% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clip_tbl.R
\name{read_clip_tbl}
\alias{read_clip_tbl}
\title{Transforms Output of Read_clip into Data.Frame}
\usage{
read_clip_tbl(x = read_clip(), ...)
}
\arguments{
\item{x}{the excel table clip to transform to data.frame}

\item{\ldots}{list of other readr options}
}
\value{
A data.frame with the contents of the clipboard.
       If the system clipboard is empty, returns NULL
}
\description{
Transforms Output of Read_clip into Data.Frame
}
\examples{
# excel contents
#' #            thingOne        thingTwo
# 1 the thing I might the thing I may
# 2    become despite coding this way

# excel contents after read_clip copies
clip_contents_from_excel_table <- c(
"thingOne\\tthingTwo",
"the thing I might\\tthe thing I may",
"become despite\\tcoding this way")

cat_in_the_hat <- read_clip_tbl(clip_contents_from_excel_table )
print(cat_in_the_hat)
#            thingOne        thingTwo
# 1 the thing I might the thing I may
# 2    become despite coding this way
}

