% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clip_tbl.R
\name{read_clip_tbl}
\alias{read_clip_tbl}
\title{Transforms Output of read_clip into dataframe.}
\usage{
read_clip_tbl(x = read_clip(), ...)
}
\arguments{
\item{x}{Defaults to reading from the clipboard, but can be substituted by a
character vector already generated by \code{\link{read_clip}}.}

\item{\ldots}{Options to pass to \code{link[readr]{read_tsv}}}
}
\value{
A data.frame with the contents of the clipboard. If the system
  clipboard is empty, returns \code{NULL}
}
\description{
Transforms clipped content into a data.frame by putting
\code{\link{read_clip}} output by using \code{link[readr]{read_tsv}}.
}
\examples{
write_clip(head(iris))
str(read_clip())
#>  chr [1:7] "Sepal.Length\\tSepal.Width\\tPetal.Length\\tPetal.Width\\tSpecies" ...

write_clip(head(iris))
str(read_clip_tbl())
#> Classes 'tbl_df', 'tbl' and 'data.frame':    6 obs. of  5 variables:
#>  $ Sepal.Length: num  5.1 4.9 4.7 4.6 5 5.4
#>  $ Sepal.Width : num  3.5 3 3.2 3.1 3.6 3.9
#>  $ Petal.Length: num  1.4 1.4 1.3 1.5 1.4 1.7
#>  $ Petal.Width : num  0.2 0.2 0.2 0.2 0.2 0.4
#>  $ Species     : chr  "setosa" "setosa" "setosa" "setosa" ...
}

