% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clip_tbl.R
\name{read_clip_tbl}
\alias{read_clip_tbl}
\title{Transforms output of \code{\link[=read_clip]{read_clip()}} into data frame.}
\usage{
read_clip_tbl(x = read_clip(), ...)
}
\arguments{
\item{x}{Defaults to reading from the clipboard, but can be substituted by a
character vector already generated by \code{\link[=read_clip]{read_clip()}}.}

\item{\ldots}{Options to pass to \code{\link[=read.table]{read.table()}}. The following
\code{\link[=read.table]{read.table()}} arguments will be passed by default, but can be
overridden by specifying them when calling \code{read_clip_tbl}: \describe{
\item{\code{header}}{\code{TRUE}}
\item{\code{sep}}{\code{"\t"}}
\item{\code{row.names}}{\code{1}}
\item{\code{stringsAsFactors}}{\code{FALSE}}
\item{\code{na.strings}}{\code{c("NA", "")}}
\item{\code{strip.white}}{\code{TRUE}} }}
}
\value{
A data frame with the contents of the clipboard. If the system
clipboard is empty, returns \code{NULL}
}
\description{
Transforms clipped content into a data frame by putting
\code{\link[=read_clip]{read_clip()}} output by using \code{\link[=read.table]{read.table()}}.
}
